function_app/requirements.txt
azure-functions
azure-ai-textanalytics
